setup.py: bump python_requires to 3.8
Since c95e4d1 (am: Fix broken guessbranch handling, 2022-03-31), b4 uses
the 'extend' action for the '--guess-branch' argument. This action is
new in Python 3.8 [1], but setup.py still lists Python 3.6 as the
minimum version.

This leads Pip to allow installing or upgrading b4 on Python 3.6 or 3.7,
but then any invocation of b4 fails with an error from the argparse
module ending with:

    ValueError: unknown action "extend"

Fix this by bumping python_requires to 3.8.

[1] https://docs.python.org/3/library/argparse.html#action

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>