ez: allow iterations to be sent in a single thread
Some projects prefer further iterations of a patch series to be sent in
the same thread as previous ones. Usually this means that the cover
letter of v2 is sent as a reply to the cover letter of v1, etc.

Add a new optional argument to get_prep_branch_as_patches, 'samethread',
defaulting to False. When True, add an 'In-Reply-To' header to the first
mail in the series, referencing the Message-ID of the previous
iterations's cover letter.

This functionality will be exposed to users in a following commit.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>